+++ code/fundamental_functions/game_functions/game_functions.h
@@ -13,7 +13,7 @@
 // play functions
 
 void cancelMoveBack(Game *g, int ind, bool left);
-void moveBackNGE(Game *g, bool autoplay, bool use_heuristique, float (*f)(Game *), memory_move_t *mem);
+void moveBackNGE(Game *g, bool autoplay, bool use_heuristique, double (*f)(Game *), memory_move_t *mem);
 void cancelAllMoveBack(Game *g, memory_move_t *mem);
 
 void lazzyMoveBack(Game *g, int indMovePawnBack, bool left);
@@ -49,7 +49,7 @@ void promotion(Game *g);
 void lienAmitie(int lig, int col, Game *g);
 
 // MOVE BACK FUNCTION
-void moveBack(Game *g, bool autoplay, bool use_heuristique, float (*f)(Game *));
+void moveBack(Game *g, bool autoplay, bool use_heuristique, double (*f)(Game *));
 /* Machine: joue à la suite les meilleurs coups en arrière */
 
 /* LIEN ENNEMITIE */
